Revert "Port DW Pose preprocessor (#1856)"
This reverts commit 0d3310f5997ed63c48971040012701817fc7249c.
