+++ .gitignore
@@ -173,4 +173,3 @@ annotator/downloads/
 # test results and expectations
 web_tests/results/
 web_tests/expectations/
-*_diff.png
