+++ annotator/openpose/body.py
@@ -11,7 +11,24 @@ from typing import NamedTuple, List, Union
 
 from . import util
 from .model import bodypose_model
-from .types import Keypoint, BodyResult
+
+class Keypoint(NamedTuple):
+    x: float
+    y: float
+    score: float = 1.0
+    id: int = -1
+
+
+class BodyResult(NamedTuple):
+    # Note: Using `Union` instead of `|` operator as the ladder is a Python
+    # 3.10 feature.
+    # Annotator code should be Python 3.8 Compatible, as controlnet repo uses
+    # Python 3.8 environment.
+    # https://github.com/lllyasviel/ControlNet/blob/d3284fcd0972c510635a4f5abe2eeb71dc0de524/environment.yaml#L6
+    keypoints: List[Union[Keypoint, None]]
+    total_score: float = 0.0
+    total_parts: int = 0
+
 
 class Body(object):
     def __init__(self, model_path):
