+++ scripts/controlnet_version.py
@@ -1,4 +1,4 @@
-version_flag = 'v1.1.237'
+version_flag = 'v1.1.234'
 
 from scripts.logging import logger
