+++ scripts/global_state.py
@@ -62,7 +62,6 @@ cn_preprocessor_modules = {
     "openpose_face": functools.partial(g_openpose_model.run_model, include_body=True, include_hand=False, include_face=True),
     "openpose_faceonly": functools.partial(g_openpose_model.run_model, include_body=False, include_hand=False, include_face=True),
     "openpose_full": functools.partial(g_openpose_model.run_model, include_body=True, include_hand=True, include_face=True),
-    "dw_openpose_full": functools.partial(g_openpose_model.run_model, include_body=True, include_hand=True, include_face=True, use_dw_pose=True),
     "clip_vision": clip,
     "color": color,
     "pidinet": pidinet,
@@ -108,7 +107,6 @@ cn_preprocessor_unloadable = {
     "openpose_hand": g_openpose_model.unload,
     "openpose_face": g_openpose_model.unload,
     "openpose_full": g_openpose_model.unload,
-    "dw_openpose_full": g_openpose_model.unload,
     "segmentation": unload_uniformer,
     "depth_zoe": unload_zoe_depth,
     "normal_bae": unload_normal_bae,
