+++ tests/annotator_tests/openpose_tests/openpose_e2e_test.py
@@ -90,29 +90,6 @@ class TestOpenposeDetector(unittest.TestCase):
             ),
             overwrite_expectation=False
         )
-
-    def test_dw(self):
-        # Need following code to pass A1111 safety check.
-        import mmengine
-        from modules import safe
-
-        def handler(module, name):
-            if module == 'mmengine.logging.history_buffer' and name in ['HistoryBuffer']:
-                return mmengine.logging.history_buffer.HistoryBuffer
-            return None
-
-        with safe.Extra(handler):
-            self.template(
-                test_image = f'{TestOpenposeDetector.image_path}/woman.jpeg',
-                expected_image = f'{TestOpenposeDetector.image_path}/expected_woman_dw_all_output.png',
-                detector_config=dict(
-                    include_body=True,
-                    include_face=True,
-                    include_hand=True,
-                    use_dw_pose=True,
-                ),
-                overwrite_expectation=False,
-            )
         
 if __name__ == '__main__':
     unittest.main()
